Makes SK2ReceiptManager an actor
- Fixes a concurrency crash that was caused by reading and writing to `sk2IntroOfferEligibility`. Made `SK2ReceiptManager` an actor to prevent this as it could potentially happen with the purchases variable too.
