+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The changelog for `SuperwallKit`. Also see the [releases](https://github.com/superwall/Superwall-iOS/releases) on GitHub.
 
+## 4.0.0-beta.4
+
+### Fixes
+
+- Fixes a crash that was caused by a concurrency issue.
+
 ## 4.0.0-beta.3
 
 ### Breaking Changes
